Remove apply
